+++ src/vs/workbench/contrib/accessibility/browser/accessibleViewContributions.ts
@@ -8,8 +8,8 @@ import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
 import { accessibleViewIsShown } from 'vs/workbench/contrib/accessibility/browser/accessibilityConfiguration';
 import * as strings from 'vs/base/common/strings';
 import { AccessibilityHelpAction, AccessibleViewAction } from 'vs/workbench/contrib/accessibility/browser/accessibleViewActions';
-import { AccessibleViewType, IAccessibleViewService } from 'vs/platform/accessibility/browser/accessibleView';
-import { AccessibleViewRegistry, IShowAccessibleViewArgs } from 'vs/platform/accessibility/browser/accessibleViewRegistry';
+import { AccessibleViewType, AdvancedContentProvider, ExtensionContentProvider, IAccessibleViewService } from 'vs/platform/accessibility/browser/accessibleView';
+import { AccessibleViewRegistry } from 'vs/platform/accessibility/browser/accessibleViewRegistry';
 import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
 
 export function descriptionForCommand(commandId: string, msg: string, noKbMsg: string, keybindingService: IKeybindingService): string {
@@ -38,9 +38,9 @@ export class AccesibleViewContributions extends Disposable {
 		super();
 		AccessibleViewRegistry.getImplementations().forEach(impl => {
 			const implementation = (accessor: ServicesAccessor) => {
-				const result: IShowAccessibleViewArgs | undefined = impl.getShowAccessibleViewArgs(accessor);
-				if (result) {
-					accessor.get(IAccessibleViewService).show(result.provider, result.position);
+				const provider: AdvancedContentProvider | ExtensionContentProvider | undefined = impl.getProvider(accessor);
+				if (provider) {
+					accessor.get(IAccessibleViewService).show(provider);
 					return true;
 				}
 				return false;
